WIP, updated star wars 8.0 project
