+++ src/ancillary-projects/starwars/starwars-api80/Properties/launchSettings.json
@@ -1,7 +1,7 @@
 ﻿{
   "$schema": "http://json.schemastore.org/launchsettings.json",
   "profiles": {
-    "StarWars: .NET 6": {
+    "StarWars: .NET 8": {
       "commandName": "Project",
       "applicationUrl": "http://localhost:5000",
       "environmentVariables": {
